Implemented addRow Servlet
On branch master
This commit addes a new servlet AddRow which facilitates adding row into
the database from the frontend. Frontend will be supporting this feature
soon.

Please hit ~/../addRow with a POST with a application/json body to add
a row into the database.

As a result after insertion, the total number of affected row, and the
total rows in the database is returned with a 200 response.

Please do note, if the serial number is even provided it would be
ignored and the backend is smart enought to calculate the latest serial
number from the last row index in the table 'winter_internship'.